chore: filter component, inscription utils, cleanup ordinals area
